Adding Allocator
Implemented a custom allocator so that new and delete are now overloaded. All STL functions inside of Memory Manager now use the custom allocator

+++ Source/MemoryAllocated.cpp
@@ -43,7 +43,7 @@
 // Public Class Functions
 //-----------------------------------------------------------------------------
 
-MemoryAllocated::MemoryAllocated(unsigned int size) :
+MemoryAllocated::MemoryAllocated(size_t size) :
                                  size(size)
 {
 }
